refactor: implement Claude's code quality improvements
- Extract duplicated formatCategory and getCategoryIcon functions to shared utility
- Improve type safety by using keyof SearchFilters instead of string for category props
- Add useCallback for hover handlers to prevent unnecessary re-renders
- Update all components to use the new shared utilities
