+++ src/components/Filters/GroupedFilterPill.tsx
@@ -1,8 +1,10 @@
-import React, { useState } from 'react';
+import React, { useState, useCallback } from 'react';
 import { X } from 'lucide-react';
+import { formatCategoryName, getCategoryIcon } from '../../utils/filterUtils';
+import type { SearchFilters } from '../../types';
 
 interface GroupedFilterPillProps {
-  category: string;
+  category: keyof SearchFilters;
   values: string[];
   // eslint-disable-next-line no-unused-vars
   onRemove: (value: string) => void;
@@ -17,50 +19,18 @@ export const GroupedFilterPill: React.FC<GroupedFilterPillProps> = ({
 }) => {
   const [isHovered, setIsHovered] = useState(false);
 
-  // Format category name for display
-  const formatCategory = (cat: string): string => {
-    const categoryNames: Record<string, string> = {
-      gradeLevels: 'Grade',
-      activityType: 'Activity',
-      seasons: 'Season',
-      thematicCategories: 'Theme',
-      culturalHeritage: 'Culture',
-      coreCompetencies: 'Competency',
-      lessonFormat: 'Format',
-      cookingMethods: 'Method',
-      academicIntegration: 'Subject',
-      socialEmotionalLearning: 'SEL',
-      location: 'Location',
-    };
-    return categoryNames[cat] || cat;
-  };
+  // Use useCallback to prevent unnecessary re-renders
+  const handleMouseEnter = useCallback(() => setIsHovered(true), []);
+  const handleMouseLeave = useCallback(() => setIsHovered(false), []);
 
-  // Get category icon
-  const getCategoryIcon = (cat: string): string => {
-    const categoryIcons: Record<string, string> = {
-      gradeLevels: '📚',
-      activityType: '🍳',
-      seasons: '🍂',
-      thematicCategories: '🌿',
-      culturalHeritage: '🌍',
-      coreCompetencies: '⭐',
-      lessonFormat: '📋',
-      cookingMethods: '🍳',
-      academicIntegration: '📚',
-      socialEmotionalLearning: '💛',
-      location: '📍',
-    };
-    return categoryIcons[cat] || '';
-  };
-
-  const formattedCategory = formatCategory(category);
+  const formattedCategory = formatCategoryName(category);
   const icon = getCategoryIcon(category);
 
   return (
     <span
       className="inline-flex items-center gap-1 px-3 py-1.5 rounded-full text-sm font-medium bg-primary-100 text-primary-800 hover:bg-primary-200 transition-colors group relative"
-      onMouseEnter={() => setIsHovered(true)}
-      onMouseLeave={() => setIsHovered(false)}
+      onMouseEnter={handleMouseEnter}
+      onMouseLeave={handleMouseLeave}
     >
       {icon && (
         <span className="text-primary-600" aria-hidden="true">
