+++ src/utils/filterUtils.ts
@@ -1,0 +1,43 @@
+/**
+ * Utility functions for filter-related operations
+ */
+
+/**
+ * Format category name for display
+ */
+export const formatCategoryName = (category: string): string => {
+  const categoryNames: Record<string, string> = {
+    gradeLevels: 'Grade',
+    activityType: 'Activity',
+    seasons: 'Season',
+    thematicCategories: 'Theme',
+    culturalHeritage: 'Culture',
+    coreCompetencies: 'Competency',
+    lessonFormat: 'Format',
+    cookingMethods: 'Method',
+    academicIntegration: 'Subject',
+    socialEmotionalLearning: 'SEL',
+    location: 'Location',
+  };
+  return categoryNames[category] || category;
+};
+
+/**
+ * Get category icon for display
+ */
+export const getCategoryIcon = (category: string): string => {
+  const categoryIcons: Record<string, string> = {
+    gradeLevels: '📚',
+    activityType: '🍳',
+    seasons: '🍂',
+    thematicCategories: '🌿',
+    culturalHeritage: '🌍',
+    coreCompetencies: '⭐',
+    lessonFormat: '📋',
+    cookingMethods: '🍳',
+    academicIntegration: '📚',
+    socialEmotionalLearning: '💛',
+    location: '📍',
+  };
+  return categoryIcons[category] || '';
+};
